Clean up code and add some docs
